[refactoring] SSE 관련 모듈 패키지 분리
